+++ expense_tracker_rest/src/main/java/com/example/expensetracker/repository/UserRepository.java
@@ -6,9 +6,9 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface UserRepository extends CrudRepository<User, Integer> {
+public interface UserRepository extends CrudRepository<User, Long> {
 
     List<User> findAll();
-    User findById(int id);
+    User findById(long id);
 
 }
